Optimized InteractionData. Updated docstrings, as well as the architecture of the documentation.
