+++ docs/cell_population.md
@@ -1,15 +1,30 @@
 # Cell Population
 
-## Base class
-
 ::: flecs.cell_population.CellPopulation
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
 
-## Test class
+## Derived classes
 
 ::: flecs.cell_population.TestCellPop
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 3
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
+
+::: flecs.cell_population.ProteinRNACellPop
+    rendering:
+        show_source: false
+        heading_level: 3
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
