+++ docs/data/composite_graphs.md
@@ -1,0 +1,13 @@
+# Composite Graphs
+
+``flecs.data.composite_graphs``
+
+::: flecs.data.composite_graphs.get_grn_string_composite_graph
+    rendering:
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_root_full_path: false
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
