+++ docs/data/protein_interactions.md
@@ -1,0 +1,13 @@
+# Protein Interactions
+
+``flecs.data.protein_interactions``
+
+::: flecs.data.protein_interactions.get_string_graph
+    rendering:
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_root_full_path: false
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
