+++ docs/data/random_graphs.md
@@ -1,9 +1,8 @@
-# Sets
+# Random Graphs
 
-This module contains two classes, ``NodeSet`` and  ``EdgeSet``, which represent sets of nodes / edges of a given type.
-These classes inherit from ``torch.nn.Module``.
+``flecs.data.random_graphs``
 
-::: flecs.sets.NodeSet
+::: flecs.data.random_graphs.get_graph_from_adj_mat
     rendering:
         show_source: false
         heading_level: 2
@@ -13,7 +12,7 @@ These classes inherit from ``torch.nn.Module``.
         show_signature: true
         members_order: source
 
-::: flecs.sets.EdgeSet
+::: flecs.data.random_graphs.get_random_adjacency_mat
     rendering:
         show_source: false
         heading_level: 2
