+++ docs/intervention.md
@@ -1,15 +1,19 @@
 # Interventions
 
-## Intervention abstract class
-
 ::: flecs.intervention.Intervention
     rendering:
-      show_source: false
-      heading_level: 3
-
-## Knockout Intervention
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
 
 ::: flecs.intervention.CrisprIntervention
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
