+++ docs/mutation.md
@@ -1,0 +1,8 @@
+::: flecs.mutation
+    rendering:
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
