+++ docs/trajectory.md
@@ -4,19 +4,31 @@
 
 :::flecs.trajectory.simulate_deterministic_trajectory_euler_steps
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
 
 ## Determistic Trajectory with solver
 
 :::flecs.trajectory.simulate_deterministic_trajectory
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
 
 ## Stochastic Trajectory
 
 :::flecs.trajectory.simulate_stochastic_trajectory
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 2
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
