+++ docs/utils.md
@@ -1,22 +1,34 @@
 # Utils
 
-## Project root
+## Production rates
 
-:::flecs.utils.get_project_root
+::: flecs.production.SimpleConv
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 3
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
 
-## Set seed
+## Decay rates
 
-:::flecs.utils.set_seed
+::: flecs.decay.exponential_decay
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 3
+        show_root_heading: true
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
 
-## Plot trajectory
+## Others
 
-:::flecs.utils.plot_trajectory
+:::flecs.utils
     rendering:
-      show_source: false
-      heading_level: 3
+        show_source: false
+        heading_level: 3
+        show_root_heading: false
+        show_signature_annotations: false
+        show_signature: true
+        members_order: source
