+++ flecs/decay.py
@@ -2,4 +2,14 @@
 
 
 def exponential_decay(obj, node_type, alpha=None):
+    """
+
+    Args:
+        obj:
+        node_type:
+        alpha:
+
+    Returns:
+
+    """
     return alpha * obj[node_type].state
