+++ flecs/mutation.py
@@ -8,11 +8,68 @@ from flecs.sets import EdgeSet
 from typing import Union
 
 ########################################################################################################################
-# Mutation Abstract class
+# Mutation functions
+########################################################################################################################
+
+
+def apply_bernoulli_mutation(obj: Union[CellPopulation, NodeSet, EdgeSet], attr_name: str, p: float, n_cells: int):
+    """
+
+    Args:
+        obj:
+        attr_name:
+        p:
+        n_cells:
+
+    Returns:
+
+    """
+
+    duplicate_attribute(obj, attr_name, n_cells)
+    noise_dist = Bernoulli(1 - p)
+
+    attr_value = obj.element_level_attr_dict[attr_name]
+
+    obj.__setattr__(attr_name, attr_value * noise_dist.sample(attr_value.shape))
+
+
+def apply_gaussian_mutation(obj: Union[CellPopulation, NodeSet, EdgeSet], attr_name: str, sigma: float, n_cells: int):
+    """
+
+    Args:
+        obj:
+        attr_name:
+        sigma:
+        n_cells:
+
+    Returns:
+
+    """
+
+    duplicate_attribute(obj, attr_name, n_cells)
+    noise_dist = Normal(0, sigma)
+
+    attr_value = obj.element_level_attr_dict[attr_name]
+
+    obj.__setattr__(attr_name, attr_value + noise_dist.sample(attr_value.shape))
+
+
+########################################################################################################################
+# Auxiliary functions
 ########################################################################################################################
 
 
 def duplicate_attribute(obj: Union[CellPopulation, NodeSet, EdgeSet], attr_name: str, n_cells: int):
+    """
+
+    Args:
+        obj:
+        attr_name:
+        n_cells:
+
+    Returns:
+
+    """
 
     assert n_cells > 1
 
@@ -26,26 +83,6 @@ def duplicate_attribute(obj: Union[CellPopulation, NodeSet, EdgeSet], attr_name:
     duplicated_attr_value = torch.cat([attr_value] * n_cells, dim=0)
 
     obj.__setattr__(attr_name, duplicated_attr_value)
-
-
-def apply_bernoulli_mutation(obj: Union[CellPopulation, NodeSet, EdgeSet], attr_name: str, p: float, n_cells: int):
-
-    duplicate_attribute(obj, attr_name, n_cells)
-    noise_dist = Bernoulli(1 - p)
-
-    attr_value = obj.element_level_attr_dict[attr_name]
-
-    obj.__setattr__(attr_name, attr_value * noise_dist.sample(attr_value.shape))
-
-
-def apply_gaussian_mutation(obj: Union[CellPopulation, NodeSet, EdgeSet], attr_name: str, sigma: float, n_cells: int):
-
-    duplicate_attribute(obj, attr_name, n_cells)
-    noise_dist = Normal(0, sigma)
-
-    attr_value = obj.element_level_attr_dict[attr_name]
-
-    obj.__setattr__(attr_name, attr_value + noise_dist.sample(attr_value.shape))
 
 
 if __name__ == "__main__":
